Move `CallContents` to `CallScreenView` file, along with `CallWebView`.
