--- features/call/src/main/kotlin/io/element/android/features/call/CallWebView.kt
@@ -1,77 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.call
-
-import android.annotation.SuppressLint
-import android.view.ViewGroup
-import android.webkit.PermissionRequest
-import android.webkit.WebChromeClient
-import android.webkit.WebView
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.viewinterop.AndroidView
-
-@Composable
-fun CallWebView(
-    url: String,
-    onPermissionsRequested: (PermissionRequest) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val isInpectionMode = LocalInspectionMode.current
-    AndroidView(
-        modifier = modifier,
-        factory = { context ->
-            WebView(context).apply {
-                if (!isInpectionMode) {
-                    setup(onPermissionsRequested)
-                    loadUrl(url)
-                }
-            }
-        },
-        update = { webView ->
-            if (!isInpectionMode) {
-                webView.loadUrl(url)
-            }
-        }
-    )
-}
-
-@SuppressLint("SetJavaScriptEnabled")
-private fun WebView.setup(onPermissionsRequested: (PermissionRequest) -> Unit) {
-    layoutParams = ViewGroup.LayoutParams(
-        ViewGroup.LayoutParams.MATCH_PARENT,
-        ViewGroup.LayoutParams.MATCH_PARENT
-    )
-
-    with(settings) {
-        javaScriptEnabled = true
-        allowContentAccess = true
-        allowFileAccess = true
-        domStorageEnabled = true
-        mediaPlaybackRequiresUserGesture = false
-        databaseEnabled = true
-        loadsImagesAutomatically = true
-        userAgentString = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36"
-    }
-
-    webChromeClient = object : WebChromeClient() {
-        override fun onPermissionRequest(request: PermissionRequest) {
-            onPermissionsRequested(request)
-        }
-    }
-}
